refactor: reorganize crate structure
