--- src/lib.rs
@@ -1,5 +1,0 @@
-mod app;
-mod terminal;
-mod ui;
-
-pub use terminal::launch;
