+++ src/viewer/error.rs
