+++ src/viewer/mod.rs
@@ -1,0 +1,12 @@
+mod app;
+mod error;
+mod keybindings;
+mod modes;
+mod utils;
+mod view;
+
+pub(crate) use crate::viewer::{
+    app::App,
+    modes::{InputMode, MainMode, Mode, NavMode, SearchMode},
+    view::View,
+};
